fix delay

+++ acceptance/acceptance_test.go
@@ -301,7 +301,7 @@ func runTest(t *testing.T, dir, coverDir string, repls testdiff.ReplacementsCont
 				require.Len(t, items, 2)
 				server.Handle(items[0], items[1], func(req testserver.Request) any {
 					if stub.DelaySeconds > 0 {
-						time.Sleep(time.Duration(stub.DelaySeconds) * time.Second)
+						time.Sleep(time.Duration(stub.DelaySeconds * float64(time.Second)))
 					}
 
 					return stub.Response
